engine/requirements.txt
# Flow AI Python Engine Dependencies
# Last Updated: 16 April 2026

# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# LLM
anthropic>=0.25.0

# Database
supabase>=2.4.0

# Config
pydantic-settings>=2.2.0
python-dotenv>=1.0.0

# HTTP Client
httpx>=0.27.0

# Google Calendar
google-auth>=2.29.0
google-api-python-client>=2.125.0

# LLM provider shim (GitHub Models local eval)
openai>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
